fix: 修正 quaternion 預設與 ONNX 輸入警告
- 修正 sim_latest_quat 語法並使用 float32
- warn when ONNX input dim mismatch & skip 無 accelerometer 模型
- DebugOverlay pass state to sync page count
- 補充 keyboard/gamepad 指令軸順序註解

Testing: python -m pytest -q

+++ state.py
@@ -79,7 +79,7 @@ class SimulationState:
     sim_latest_action_raw: np.ndarray = field(default_factory=lambda: np.zeros(12))
     sim_latest_final_ctrl: np.ndarray = field(default_factory=lambda: np.zeros(12))
     sim_latest_pos: np.ndarray = field(default_factory=lambda: np.zeros(3))
-    sim_latest_quat: np.ndarray = field(default_factory=lambda: np.array([1.,0.,0.,0.]))
+    sim_latest_quat: np.ndarray = field(default_factory=lambda: np.array([1., 0., 0., 0.], dtype=np.float32))
     sim_latest_joint_positions: np.ndarray = field(default_factory=lambda: np.zeros(12))
 
     # 使用者命令與調校參數
